HW-7 Retrofit 관련 부분 분리 및 Inject
