+++ class14/myung777/StudyFork/app/src/main/java/com/deepco/studyfork/api/RetrofitHelper.kt
@@ -1,0 +1,48 @@
+package com.deepco.studyfork.api
+
+import com.deepco.studyfork.BuildConfig
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Inject
+
+class RetrofitHelper @Inject constructor() {
+
+    fun create(): RetrofitService {
+        val httpLoggingInterceptor = HttpLoggingInterceptor()
+        if (BuildConfig.DEBUG) {
+            httpLoggingInterceptor.level = HttpLoggingInterceptor.Level.BODY
+        } else {
+            httpLoggingInterceptor.level = HttpLoggingInterceptor.Level.NONE
+        }
+
+        val headerInterceptor = Interceptor {
+            val request = it.request()
+                .newBuilder()
+                .addHeader("X-Naver-Client-Id", CLIENT_ID)
+                .addHeader("X-Naver-Client-Secret", CLIENT_SECRET)
+                .build()
+            return@Interceptor it.proceed(request)
+        }
+
+        val client = OkHttpClient.Builder()
+            .addInterceptor(headerInterceptor)
+            .addInterceptor(httpLoggingInterceptor)
+            .build()
+
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL_NAVER_API)
+            .client(client)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(RetrofitService::class.java)
+    }
+
+    companion object {
+        private const val BASE_URL_NAVER_API = "https://openapi.naver.com/"
+        private const val CLIENT_ID = "hNSL7xP8VekVAD6qbXn4"
+        private const val CLIENT_SECRET = "PAzBOvOWow"
+    }
+}
